+++ examples/wide-string.rs
@@ -40,11 +40,11 @@ pub fn main_routine() -> Result<(), NCurseswError> {
     let str1 = "\u{41f}\u{440}\u{438}\u{432}\u{435}\u{442} is hello in russian!\n";
     let str2 = "🙈🙊🙉🙈🙊🙉\n";
 
-    let wide_string1 = WideString::from_str(str1);
-    let wide_string2 = WideString::from_str(str2);
+    let wide_string1 = &WideString::from_str(str1);
+    let wide_string2 = &WideString::from_str(str2);
 
-    addwstr(&wide_string1)?;
-    addwstr(&wide_string2)?;
+    addwstr(wide_string1)?;
+    addwstr(wide_string2)?;
 
     refresh()?;